big refactor to create plotdata
