+++ src/scores/plotdata/__init__.py
@@ -1,0 +1,14 @@
+"""
+Import the functions from the implementations into the public API
+"""
+
+from scores.plotdata.murphy_impl import murphy_score, murphy_thetas
+from scores.plotdata.qq_impl import qq
+from scores.plotdata.roc_impl import roc
+
+__all__ = [
+    "murphy_score",
+    "murphy_thetas",
+    "qq",
+    "roc",
+]
